Bare Minimum
I've updated the formatting so it looks relatively pretty. I also added
a footer with a apology/disclaimer just in case the TA's get a bad joke
while grading the project. I suppose I could turn it in now because it
meets the requirements, but I'll probably spend some more time finding
ways to make it more pretty and fun.